+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## master
 
+## 1.2.0-3 (beta)
+- Introduce `melody-streams` API [#102](https://github.com/trivago/melody/pull/102)
+
 ## 1.2.0-2 (beta)
 
 ### Features
@@ -11,7 +14,6 @@
 - Introduce `useAtom` hook [#79](https://github.com/trivago/melody/pulls/79)
 - Introduce `useStore` hook and performance marks for hooks [#98](https://github.com/trivago/melody/pulls/98)
 - Added async mounting of components [#82](https://github.com/trivago/melody/pull/82)
-- Introduce `melody-streams` API [#102](https://github.com/trivago/melody/pull/102)
 
 ### Fixes